Use script to launch test containers
There is now a little bit of setup with environment variables and
temporary directories that make it easier (necessary) to run things in
scripts

This might make the code more difficult to maintain since new developers
won't be to happy to see all the code. A way to re-factor
this is to use docker-compose to reduce the complexity. For now though,
this solution works, and has benefit of being 1) being simpler by
avoiding an extra framework (docker-compose) and 2) the setup is closer to a docker-less
solution, which I argue also makes it easier (i.e. if you want to run
the tests without docker, the code "just works", you don't _have_ to use
docker)