Bugfix: allow normal iterators in `feature_extraction`
To get the progress bar, previous code assumed that the iterator would
have a `read_kwargs` available. This is not the case for normal
iterators, only `TiffPrefetch`.

We not support a normal progress bar for TiffPrefetch and a live
progress bar with others that will update a bar but will not now how
many entries there are