Add errorhandling in case tritonserver returns failures
This may not be desired behaviour: if all tiles fail, the program now
continues execution. This could be frustrating, e.g. if the server has
some downtime during a large workload. However, the upside is that if you have
one or two bad slides in a large dataset, the whole pipeline doesn't crash.

Pick your poison...